+++ zenoh/src/net/runtime/mod.rs
@@ -36,10 +36,11 @@ use futures::{stream::StreamExt, Future};
 use tokio::task::JoinHandle;
 use tokio_util::sync::CancellationToken;
 use uhlc::{HLCBuilder, HLC};
+use zenoh_config::wrappers::ZenohId;
 use zenoh_link::{EndPoint, Link};
 use zenoh_plugin_trait::{PluginStartArgs, StructVersion};
 use zenoh_protocol::{
-    core::{Locator, WhatAmI, ZenohId},
+    core::{Locator, WhatAmI},
     network::NetworkMessage,
 };
 use zenoh_result::{bail, ZResult};
